Moved the Rules/SRLArr back to the core
