+++ config-example.properties
@@ -123,4 +123,4 @@ graph.dataaccess=RAM_STORE
 
 # Spatial Rules require you to provide Polygons in which the rules are enforced
 # Below line contains the default location for these rules
-# spatial_rules.location=web/src/main/resources/com/graphhopper/http/countries.geo.json
+# spatial_rules.location=web/src/main/resources/com/graphhopper/spatialrules/countries.geo.json
