+++ core/src/main/java/com/graphhopper/routing/util/spatialrules/SpatialRuleLookupArray.java
@@ -15,11 +15,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package com.graphhopper.srlarray;
+package com.graphhopper.routing.util.spatialrules;
 
-import com.graphhopper.routing.util.spatialrules.Polygon;
-import com.graphhopper.routing.util.spatialrules.SpatialRule;
-import com.graphhopper.routing.util.spatialrules.SpatialRuleLookup;
 import com.graphhopper.util.shapes.BBox;
 import com.graphhopper.util.shapes.GHPoint;
 
@@ -31,7 +28,7 @@ import java.util.*;
  *
  * @author Robin Boldt
  */
-class SpatialRuleLookupArray implements SpatialRuleLookup {
+public class SpatialRuleLookupArray implements SpatialRuleLookup {
 
     // resolution in full decimal degrees
     private final double resolution;
@@ -53,7 +50,7 @@ class SpatialRuleLookupArray implements SpatialRuleLookup {
      * @param exact      if exact it will also perform a polygon contains for border tiles, might fail for small holes
      *                   in the Polygon that are not represented in the tile array.
      */
-    SpatialRuleLookupArray(List<SpatialRule> spatialRules, double resolution, boolean exact) {
+    public SpatialRuleLookupArray(List<SpatialRule> spatialRules, double resolution, boolean exact) {
         bounds = BBox.createInverse(false);
         for (SpatialRule spatialRule : spatialRules) {
             for (Polygon polygon : spatialRule.getBorders()) {
