+++ web/src/main/java/com/graphhopper/http/DefaultModule.java
@@ -20,8 +20,8 @@ package com.graphhopper.http;
 import com.google.inject.AbstractModule;
 import com.google.inject.name.Names;
 import com.graphhopper.GraphHopper;
-import com.graphhopper.srlarray.SpatialRuleLookupBuilder;
-import com.graphhopper.countries.CountriesSpatialRuleFactory;
+import com.graphhopper.spatialrules.SpatialRuleLookupBuilder;
+import com.graphhopper.spatialrules.CountriesSpatialRuleFactory;
 import com.graphhopper.json.GHJson;
 import com.graphhopper.json.GHJsonBuilder;
 import com.graphhopper.json.geo.JsonFeatureCollection;
