+++ web/src/main/resources/com/graphhopper/spatialrules/README.MD
