+++ web/src/main/resources/com/graphhopper/spatialrules/countries.geo.json
